[FIX] l10n_ar_ux: monkey patch fiscal_position post_load
Task: 34656
