+++ l10n_ar_ux/__init__.py
@@ -4,15 +4,16 @@
 ##############################################################################
 from . import controllers
 from . import models
+from odoo import api
 from . import reports
 from . import wizards
 from .hooks import post_init_hook
-
 from odoo.addons.l10n_latam_invoice_document.models.account_move import AccountMove
+from odoo.addons.l10n_ar.models.account_fiscal_position import AccountFiscalPosition
 from odoo.exceptions import UserError
 
 
-def monkey_patch_inverse_l10n_latam_document_number():
+def monkey_patches():
     # monkey patch
     orginal_method = AccountMove._inverse_l10n_latam_document_number
 
@@ -41,3 +42,14 @@ def monkey_patch_inverse_l10n_latam_document_number():
                                       ' the POS number if there is not posted (or posted before) invoices'))
 
     AccountMove._inverse_l10n_latam_document_number = _inverse_l10n_latam_document_number
+
+    # monkey patch
+    @api.model
+    def _get_fiscal_position(self, partner, delivery=None):
+        if self.env.company.country_id.code == "AR":
+            self = self.with_context(
+                company_code='AR',
+                l10n_ar_afip_responsibility_type_id=partner.l10n_ar_afip_responsibility_type_id.id)
+        return super(AccountFiscalPosition, self)._get_fiscal_position(partner, delivery=delivery)
+
+    AccountFiscalPosition._get_fiscal_position = _get_fiscal_position
